level

+++ project/athena/src/com/athena/level/LevelBean.java
@@ -1,18 +1,20 @@
 package com.athena.level;
 
+import java.math.BigDecimal;
+
 public class LevelBean {
 	private int level;
-	private float discount;
+	private double discount;
 	public int getLevel() {
 		return level;
 	}
 	public void setLevel(int level) {
 		this.level = level;
 	}
-	public float getDiscount() {
+	public double getDiscount() {
 		return discount;
 	}
-	public void setDiscount(float discount) {
+	public void setDiscount(double discount) {
 		this.discount = discount;
 	}
