+++ project/database/creater.sql
@@ -3,6 +3,6 @@ CREATE DATABASE athena;
 USE athena;
 CREATE TABLE member (
   `level` INT(11) NOT NULL,
-  `discount` FLOAT(6) DEFAULT 1,
+  `discount` DOUBLE(3,2) DEFAULT 1,
   PRIMARY KEY (`level`)
 )
